kafka-situation-docker/detector/requirements.txt
kafka-python
sklearn
nltk
numpy